+++ packages/voice/src/models/v2/index.ts
@@ -35,5 +35,7 @@ export * from './batch-details';
 export * from './money';
 export * from './call';
 export * from './session';
+export * from './webhook-request';
+export * from './webhook-response';
 export * from './enums';
 export * from './requests';