+++ packages/voice/src/rest/v2/index.ts
@@ -3,3 +3,8 @@ export * from './voice-v2-service';
 export { CallsApi as VoiceV2CallsApi, CallsApiFixture as VoiceV2CallsApiFixture } from './calls';
 export { BatchesApi as VoiceV2BatchesApi, BatchesApiFixture as VoiceV2BatchesApiFixture } from './batches';
 export { SessionsApi as VoiceV2SessionsApi, SessionsApiFixture as VoiceV2SessionsApiFixture } from './sessions';
+export {
+  SinchEvents as VoiceV2SinchEvents,
+  SinchEventsFixture as VoiceV2SinchEventsFixture,
+} from './sinch-events';
+export type { VoiceV2SinchEventsParameters } from './sinch-events';