+++ packages/voice/src/rest/v2/sinch-events/index.ts
@@ -1,0 +1,2 @@
+export * from './sinch-events';
+export * from './sinch-events.jest.fixture';